Loan Risk Prediction Using Transaction Information
Aim/Objective of the project:

The objective of proposed work is to predict loan credit risk and determine the probability of non-payment of bank financial services example whether a person will pay back a loan or not. The other objective of the project is to study the ability of neural network algorithms to handle the problem of predicting credit default that measures the creditworthiness of the loan application over a time period. However, there are many risks related to bank loans, for the bank and for those who get the loans. Risk prediction and monitoring is critical for the success of the business model. Credit risk is the probability that a customer won’t be able to make a required payment, causing a loss for the bank or financial institute that provided the loan.